Avoid exception if redmine_organizations plugin is not installed (fixes #1)
The redmine_organizations plugin adds a `users.organization_id` column
to store the organization (= main group) of each user. This can enhance
this plugin features by adding a "visibility" scope for your reminders.
That being said, the plugin shouldn't break if redmine_organizations is
not here, but I forgot one occurrence of User.current.organization_id in
the model. Note that the tests were already breaking on a fresh
database structure.